+++ modules/capitalize.rb
@@ -1,0 +1,8 @@
+require_relative "./decorator"
+
+class CapitalizeDecorator < BaseDecorator
+    def correct_name
+      @nameable.correct_name.capitalize
+    end
+  end
+
